Moved asteroidType and ASteroid to CommonAsteroid
