+++ AsteroidsNetbeansModules/Asteroid/src/main/java/dk/sdu/mmmi/cbse/asteroid/AsteroidSplitter.java
@@ -5,8 +5,9 @@
  */
 package dk.sdu.mmmi.cbse.asteroid;
 
-import static dk.sdu.mmmi.cbse.asteroid.AsteroidType.MEDIUM;
-import static dk.sdu.mmmi.cbse.asteroid.AsteroidType.SMALL;
+import dk.sdu.mmmi.cbse.common.asteroids.Asteroid;
+import static dk.sdu.mmmi.cbse.common.asteroids.AsteroidType.MEDIUM;
+import static dk.sdu.mmmi.cbse.common.asteroids.AsteroidType.SMALL;
 import dk.sdu.mmmi.cbse.common.data.Entity;
 import dk.sdu.mmmi.cbse.common.data.GameData;
 import dk.sdu.mmmi.cbse.common.data.World;
