+++ AsteroidsNetbeansModules/CommonAsteroids/src/main/java/dk/sdu/mmmi/cbse/common/asteroids/Asteroid.java
@@ -4,7 +4,17 @@ import dk.sdu.mmmi.cbse.common.data.Entity;
 
 /**
  *
- * @author corfixen
+ * @author Phillip Olsen
  */
 public class Asteroid extends Entity {
+
+    private AsteroidType type;
+
+    public Asteroid(AsteroidType type) {
+        this.type = type;
+    }
+
+    public String getSize() {
+        return type.getSize();
+    }
 }
